Get disemvowel mostly working for small inputs.
It currently doesn't pass the stdin stout test because it returns a bunch of null characters
for any part of the buffer size that doesn't have text written to it.

Going forward we need to put our read and write statements in a while loop, so we can process more than 1 buffer
full of text from a file. As it stands, we would be unable to process really long files.

We also need to process files provided in the command line as well, but that will come later.