Begin work on summarize_tree using stat().
We've gotten some of the logic for things like is_dir probably correct, but we were confused about how to actually
use the extra process_path() function that was provided above process_directory. Because of this we don't really
recurse through things properly just yet.

As of right now the program compiles and not much else.