Continue working on summarize_tree.c
We've got recursion going now, but it breaks things. For a while we were getting wrong outputs, but now we are getting none.

Probably gonna want to check our conditional statements and error handling for stat to make sure our program isn't dying there
or ending up in an infinite loop because our strcmp is funky in some way.